智能体项目/requirements.txt
python-dotenv>=1.0.0
rich>=13.7.0
numpy>=1.24.0
pydantic>=2.7.0
openai>=1.51.0
duckduckgo-search>=6.2.10